refactor: split base and exprs into multiple files
